parameterization of the tree
