+++ src/FromTo.java
@@ -1,4 +1,4 @@
 public interface FromTo {
-   <T extends Human> void saveTo(GenealogicalTree<T> family);
+   <T extends TreeElement> void saveTo(GenealogicalTree<T> family);
    GenealogicalTree readFrom(String fileName);
 }
