+++ src/TreeElement.java
@@ -1,3 +1,26 @@
-public abstract class TreeElement {
+import java.io.Serializable;
 
+public abstract class TreeElement implements Serializable, Comparable<TreeElement> {
+     String name;
+
+     int birthday;
+     int deathDay;
+
+    public TreeElement(String name, int birthday, int deathDay) {
+        this.name = name;
+        this.birthday = birthday;
+        this.deathDay = deathDay;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getBirthday() {
+        return birthday;
+    }
+
+    public int getDeathDay() {
+        return deathDay;
+    }
 }
